unify powershell and spawn code into a win32-specific location
